extract mdnc from main
